+++ src/beatmap/shared/difficulty.ts
@@ -7,6 +7,7 @@ export const DifficultyRanking: { [key in DifficultyName]: DifficultyRank } = {
    Hard: 5,
    Expert: 7,
    ExpertPlus: 9,
+   'Expert+': 9,
 } as const;
 
 /** Difficulty rename to human readable. */
@@ -16,4 +17,5 @@ export const DifficultyRename: { [key in DifficultyName]: string } = {
    Hard: 'Hard',
    Expert: 'Expert',
    ExpertPlus: 'Expert+',
+   'Expert+': 'Expert+',
 } as const;