+++ src/types/beatmap/shared/difficulty.ts
@@ -1,5 +1,5 @@
 /** Available difficulty from base game. */
-export type DifficultyName = 'Easy' | 'Normal' | 'Hard' | 'Expert' | 'ExpertPlus';
+export type DifficultyName = 'Easy' | 'Normal' | 'Hard' | 'Expert' | 'ExpertPlus' | 'Expert+';
 
 /** Difficulty rename to human readable. */
 export type DifficultyRank = 1 | 3 | 5 | 7 | 9;